Updated CheckVersion()
Modify update prompt to proceed only if user clicks "OK"

This change ensures that the update process will only proceed if the user explicitly clicks "OK" in the message box. If the user clicks "Cancel" or exits the message box in any other way, they will not be redirected to the update.